+++ src/clientHash.h
@@ -5,14 +5,14 @@
 
 typedef struct {
     unsigned long clientIpPort;
-    unsigned int serverPort;
+    unsigned short serverPort;
     UT_hash_handle hh;
 } CLIENT, *PCLIENT;
 
-unsigned long combineIpPort(unsigned int ip, unsigned int port);
-void clientAdd(unsigned int clientIp, unsigned int clientPort, unsigned int serverPort);
-PCLIENT clientFind(unsigned int clientIp, unsigned int clientPort);
-void clientDelete(unsigned int clientIp, unsigned int clientPort);
+unsigned long combineIpPort(unsigned int ip, unsigned short port);
+void clientAdd(unsigned int clientIp, unsigned short clientPort, unsigned short serverPort);
+PCLIENT clientFind(unsigned int clientIp, unsigned short clientPort);
+void clientDelete(unsigned int clientIp, unsigned short clientPort);
 void clientDeleteAll();
 void clientPrint();
 int clientSort(PCLIENT a, PCLIENT b);
