+++ src/sharedLibrary.h
@@ -23,12 +23,12 @@ typedef struct
 
 /* Function Prototypes */
 void systemFatal(const char* message);
-unsigned int addRuleToMaps(unsigned int clientIp, unsigned int clientPort);
-void deleteRuleFromMap(unsigned int clientIp, unsigned int clientPort, unsigned int serverPort);
-int cliFind(unsigned int clientIp, unsigned int clientPort, unsigned int *srvPort);
-int srvFind(unsigned int serverPort, unsigned int *clientIp, unsigned int *clientPort);
-int rlFind(unsigned int clientDestPort, unsigned int *serverDestPort, unsigned int *serverDestIp);
-int rlAdd(unsigned int clientDestPort, unsigned int serverDestPort, unsigned int serverDestIp);
+unsigned short addRuleToMaps(unsigned int clientIp, unsigned short clientPort);
+void deleteRuleFromMap(unsigned int clientIp, unsigned short clientPort, unsigned short serverPort);
+int cliFind(unsigned int clientIp, unsigned short clientPort, unsigned short *srvPort);
+unsigned short srvFind(unsigned short serverPort, unsigned int *clientIp, unsigned short *clientPort);
+int rlFind(unsigned short clientDestPort, unsigned short *serverDestPort, unsigned int *serverDestIp);
+int rlAdd(unsigned short clientDestPort, unsigned short serverDestPort, unsigned int serverDestIp);
 char* rlToStr();
